backend/requirements.txt
Django==3.2.10
djangorestframework==3.12.4
django-filter==21.1
drf-extra-fields==3.2.1
pillow==8.4.0

djoser==2.1.0
chardet==4.0.0

gunicorn==20.1.0
